[Parley] refactor: Split ScriptParameterUIManager.cs (#706)
Split 666-line file into focused services:
- ScriptParameterUIManager.cs (269 lines) - core row management
- ParameterValidationService.cs (244 lines) - duplicate key validation
- ParameterPersistenceService.cs (233 lines) - UI-to-model sync

Tests: All 500 tests pass ✅
