+++ Parley/Parley/Services/ScriptParameterUIManager.cs
@@ -1,15 +1,10 @@
 using Avalonia;
 using Avalonia.Controls;
-using Avalonia.Interactivity;
 using Avalonia.Media;
-using Avalonia.Styling;
 using DialogEditor.Models;
 using DialogEditor.Utils;
 using System;
 using Radoub.Formats.Logging;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace DialogEditor.Services
 {
@@ -17,6 +12,9 @@ namespace DialogEditor.Services
     /// Manages script parameter UI components and synchronization with dialog model.
     /// Handles both conditional (DialogPtr) and action (DialogNode) parameters.
     /// Extracted from MainWindow.axaml.cs to reduce duplication and improve testability.
+    ///
+    /// Refactored in #706: Validation logic extracted to ParameterValidationService,
+    /// persistence logic extracted to ParameterPersistenceService.
     /// </summary>
     public class ScriptParameterUIManager
     {
@@ -25,6 +23,9 @@ namespace DialogEditor.Services
         private readonly Action _triggerAutoSave;
         private readonly Func<bool> _isPopulatingProperties;
         private readonly Func<TreeViewSafeNode?> _getSelectedNode;
+
+        private readonly ParameterValidationService _validationService;
+        private readonly ParameterPersistenceService _persistenceService;
 
         public ScriptParameterUIManager(
             Func<string, Control?> findControl,
@@ -38,6 +39,9 @@ namespace DialogEditor.Services
             _triggerAutoSave = triggerAutoSave ?? throw new ArgumentNullException(nameof(triggerAutoSave));
             _isPopulatingProperties = isPopulatingProperties ?? throw new ArgumentNullException(nameof(isPopulatingProperties));
             _getSelectedNode = getSelectedNode ?? throw new ArgumentNullException(nameof(getSelectedNode));
+
+            _validationService = new ParameterValidationService(setStatusMessage);
+            _persistenceService = new ParameterPersistenceService(findControl, _validationService);
         }
 
         /// <summary>
@@ -108,11 +112,11 @@ namespace DialogEditor.Services
             keyTextBox.TextChanged += (s, e) =>
             {
                 // Re-validate on every change to clear red border when duplicate is resolved
-                ValidateDuplicateKeys(parent, keyTextBox, isCondition);
+                _validationService.ValidateDuplicateKeys(parent, keyTextBox, isCondition);
             };
             keyTextBox.LostFocus += (s, e) =>
             {
-                ValidateDuplicateKeys(parent, keyTextBox, isCondition);
+                _validationService.ValidateDuplicateKeys(parent, keyTextBox, isCondition);
                 OnParameterChanged(isCondition, _getSelectedNode());
             };
             grid.Children.Add(keyTextBox);
@@ -174,193 +178,6 @@ namespace DialogEditor.Services
         }
 
         /// <summary>
-        /// Validates that a key is not duplicated in the parameter panel.
-        /// Issue #287: Prevents duplicate keys which would cause data loss.
-        /// Red border stays until the duplicate is corrected.
-        /// Issue #141: Fixed to clear ALL red borders when duplicate is resolved.
-        /// </summary>
-        private void ValidateDuplicateKeys(StackPanel parent, TextBox currentKeyTextBox, bool isCondition)
-        {
-            string currentKey = currentKeyTextBox.Text?.Trim() ?? "";
-
-            // If key is empty, clear any warning state and revalidate all to clear orphaned red borders
-            if (string.IsNullOrWhiteSpace(currentKey))
-            {
-                ClearDuplicateWarning(currentKeyTextBox);
-                RevalidateAllKeys(parent, isCondition);
-                return;
-            }
-
-            int duplicateCount = 0;
-            var allKeyTextBoxes = new List<TextBox>();
-
-            foreach (var child in parent.Children)
-            {
-                if (child is Grid paramGrid)
-                {
-                    var textBoxes = paramGrid.Children.OfType<TextBox>().ToList();
-                    if (textBoxes.Count >= 1)
-                    {
-                        var keyTextBox = textBoxes[0];
-                        string key = keyTextBox.Text?.Trim() ?? "";
-
-                        if (key.Equals(currentKey, StringComparison.Ordinal))
-                        {
-                            duplicateCount++;
-                            allKeyTextBoxes.Add(keyTextBox);
-                        }
-                    }
-                }
-            }
-
-            if (duplicateCount > 1)
-            {
-                // Show warning - duplicate key detected (use theme error color for accessibility)
-                var errorBrush = GetErrorBrush();
-                currentKeyTextBox.BorderBrush = errorBrush;
-                currentKeyTextBox.BorderThickness = new Thickness(2);
-
-                // Also mark all other textboxes with the same key
-                foreach (var tb in allKeyTextBoxes)
-                {
-                    tb.BorderBrush = errorBrush;
-                    tb.BorderThickness = new Thickness(2);
-                }
-
-                _setStatusMessage($"⚠️ Duplicate key '{currentKey}' - only one value will be saved!");
-
-                UnifiedLogger.LogApplication(LogLevel.WARN,
-                    $"Duplicate key detected: '{currentKey}' appears {duplicateCount} times in {(isCondition ? "condition" : "action")} parameters");
-            }
-            else
-            {
-                // No duplicate for current key - clear warning on this textbox
-                ClearDuplicateWarning(currentKeyTextBox);
-
-                // Issue #141: Revalidate ALL keys to clear orphaned red borders
-                // When a duplicate is resolved by editing one key, the other key's
-                // textbox still has a red border that needs to be cleared
-                RevalidateAllKeys(parent, isCondition);
-            }
-        }
-
-        /// <summary>
-        /// Revalidates all key textboxes in the panel to clear orphaned red borders.
-        /// Issue #141: Called when a key changes to ensure previously-duplicate keys get cleared.
-        /// </summary>
-        private void RevalidateAllKeys(StackPanel parent, bool isCondition)
-        {
-            // Build a count of each key to find actual duplicates
-            var keyCounts = new Dictionary<string, int>(StringComparer.Ordinal);
-            var keyTextBoxes = new Dictionary<string, List<TextBox>>(StringComparer.Ordinal);
-
-            foreach (var child in parent.Children)
-            {
-                if (child is Grid paramGrid)
-                {
-                    var textBoxes = paramGrid.Children.OfType<TextBox>().ToList();
-                    if (textBoxes.Count >= 1)
-                    {
-                        var keyTextBox = textBoxes[0];
-                        string key = keyTextBox.Text?.Trim() ?? "";
-
-                        if (!string.IsNullOrWhiteSpace(key))
-                        {
-                            if (!keyCounts.ContainsKey(key))
-                            {
-                                keyCounts[key] = 0;
-                                keyTextBoxes[key] = new List<TextBox>();
-                            }
-                            keyCounts[key]++;
-                            keyTextBoxes[key].Add(keyTextBox);
-                        }
-                        else
-                        {
-                            // Empty key - clear any red border
-                            ClearDuplicateWarning(keyTextBox);
-                        }
-                    }
-                }
-            }
-
-            // Now update visual state for each key
-            // Get theme error brush once for efficiency
-            var errorBrush = GetErrorBrush();
-
-            foreach (var kvp in keyCounts)
-            {
-                string key = kvp.Key;
-                int count = kvp.Value;
-
-                if (count > 1)
-                {
-                    // Still a duplicate - use theme error color
-                    foreach (var tb in keyTextBoxes[key])
-                    {
-                        tb.BorderBrush = errorBrush;
-                        tb.BorderThickness = new Thickness(2);
-                    }
-                }
-                else
-                {
-                    // Not a duplicate (anymore) - clear error border
-                    foreach (var tb in keyTextBoxes[key])
-                    {
-                        ClearDuplicateWarning(tb);
-                    }
-                }
-            }
-        }
-
-        /// <summary>
-        /// Gets the theme-aware error brush for validation errors.
-        /// Falls back to red if theme error color is not available.
-        /// Issue #141: Uses theme colors for colorblind accessibility.
-        /// </summary>
-        private IBrush GetErrorBrush()
-        {
-            var app = Application.Current;
-            if (app?.Resources.TryGetResource("ThemeError", ThemeVariant.Default, out var errorBrush) == true
-                && errorBrush is IBrush brush)
-            {
-                return brush;
-            }
-            // Fallback to standard red
-            return Brushes.Red;
-        }
-
-        /// <summary>
-        /// Gets the theme-aware success brush for validation success feedback.
-        /// Falls back to green if theme success color is not available.
-        /// Issue #141: Uses theme colors for colorblind accessibility.
-        /// </summary>
-        private IBrush GetSuccessBrush()
-        {
-            var app = Application.Current;
-            if (app?.Resources.TryGetResource("ThemeSuccess", ThemeVariant.Default, out var successBrush) == true
-                && successBrush is IBrush brush)
-            {
-                return brush;
-            }
-            // Fallback to standard green
-            return Brushes.LightGreen;
-        }
-
-        /// <summary>
-        /// Clears the duplicate key warning visual state from a TextBox.
-        /// </summary>
-        private void ClearDuplicateWarning(TextBox textBox)
-        {
-            // Only clear if currently showing error border (duplicate warning)
-            // Check for both theme error brush and fallback red
-            if (textBox.BorderThickness.Top >= 2)
-            {
-                textBox.BorderBrush = null; // Reset to default theme
-                textBox.BorderThickness = new Thickness(1);
-            }
-        }
-
-        /// <summary>
         /// Issue #289: Checks if either parameter panel (conditions or actions) has duplicate keys.
         /// Called by MainWindow before saving to prevent data corruption.
         /// </summary>
@@ -369,42 +186,12 @@ namespace DialogEditor.Services
             var conditionsPanel = _findControl("ConditionsParametersPanel") as StackPanel;
             var actionsPanel = _findControl("ActionsParametersPanel") as StackPanel;
 
-            if (conditionsPanel != null && HasDuplicateKeys(conditionsPanel))
+            if (conditionsPanel != null && _validationService.HasDuplicateKeys(conditionsPanel))
                 return true;
 
-            if (actionsPanel != null && HasDuplicateKeys(actionsPanel))
+            if (actionsPanel != null && _validationService.HasDuplicateKeys(actionsPanel))
                 return true;
 
-            return false;
-        }
-
-        /// <summary>
-        /// Checks if a parameter panel has any duplicate keys.
-        /// Issue #287: Used to block saving when duplicates exist.
-        /// </summary>
-        private bool HasDuplicateKeys(StackPanel panel)
-        {
-            var keys = new HashSet<string>(StringComparer.Ordinal);
-
-            foreach (var child in panel.Children)
-            {
-                if (child is Grid paramGrid)
-                {
-                    var textBoxes = paramGrid.Children.OfType<TextBox>().ToList();
-                    if (textBoxes.Count >= 1)
-                    {
-                        string key = textBoxes[0].Text?.Trim() ?? "";
-                        if (!string.IsNullOrWhiteSpace(key))
-                        {
-                            if (!keys.Add(key))
-                            {
-                                // Duplicate found
-                                return true;
-                            }
-                        }
-                    }
-                }
-            }
             return false;
         }
 
@@ -427,7 +214,7 @@ namespace DialogEditor.Services
             // Issue #287: Check for duplicate keys BEFORE saving - block save if duplicates exist
             var panelName = isCondition ? "ConditionsParametersPanel" : "ActionsParametersPanel";
             var panel = _findControl(panelName) as StackPanel;
-            if (panel != null && HasDuplicateKeys(panel))
+            if (panel != null && _validationService.HasDuplicateKeys(panel))
             {
                 UnifiedLogger.LogApplication(LogLevel.WARN,
                     $"OnParameterChanged: BLOCKED - Duplicate keys detected in {panelName}. Parameters NOT saved.");
@@ -443,7 +230,7 @@ namespace DialogEditor.Services
                 // Conditional parameters are on the DialogPtr
                 if (sourcePtr != null)
                 {
-                    UpdateConditionParamsFromUI(sourcePtr);
+                    _persistenceService.UpdateConditionParamsFromUI(sourcePtr);
                     _setStatusMessage("Condition parameters updated");
                 }
                 else
@@ -454,7 +241,7 @@ namespace DialogEditor.Services
             else
             {
                 // Action parameters are on the DialogNode
-                UpdateActionParamsFromUI(dialogNode);
+                _persistenceService.UpdateActionParamsFromUI(dialogNode);
                 _setStatusMessage("Action parameters updated");
             }
 
@@ -462,205 +249,21 @@ namespace DialogEditor.Services
         }
 
         /// <summary>
-        /// Updates conditional parameters from UI to model
+        /// Updates conditional parameters from UI to model.
+        /// Delegated to ParameterPersistenceService.
         /// </summary>
         public void UpdateConditionParamsFromUI(DialogPtr ptr, string? scriptName = null)
         {
-            UnifiedLogger.LogApplication(LogLevel.DEBUG,
-                $"UpdateConditionParamsFromUI: ENTRY - ptr has {ptr.ConditionParams.Count} existing params");
-
-            ptr.ConditionParams.Clear();
-            var panel = _findControl("ConditionsParametersPanel") as StackPanel;
-            if (panel == null)
-            {
-                UnifiedLogger.LogApplication(LogLevel.ERROR,
-                    "UpdateConditionParamsFromUI: ConditionsParametersPanel NOT FOUND - parameters will be empty!");
-                return;
-            }
-
-            // Get script name from UI if not provided
-            if (string.IsNullOrEmpty(scriptName))
-            {
-                var scriptTextBox = _findControl("ScriptAppearsTextBox") as TextBox;
-                scriptName = scriptTextBox?.Text;
-            }
-
-            UnifiedLogger.LogApplication(LogLevel.DEBUG,
-                $"UpdateConditionParamsFromUI: Found ConditionsParametersPanel with {panel.Children.Count} children");
-
-            var autoTrimCheckBox = _findControl("AutoTrimConditionsCheckBox") as CheckBox;
-            bool autoTrim = autoTrimCheckBox?.IsChecked ?? true;
-
-            UnifiedLogger.LogApplication(LogLevel.DEBUG, $"UpdateConditionParamsFromUI: Auto-Trim = {autoTrim}");
-
-            ProcessParameterPanel(panel, ptr.ConditionParams, autoTrim, scriptName);
-
-            UnifiedLogger.LogApplication(LogLevel.DEBUG,
-                $"UpdateConditionParamsFromUI: EXIT - ptr now has {ptr.ConditionParams.Count} params");
+            _persistenceService.UpdateConditionParamsFromUI(ptr, scriptName);
         }
 
         /// <summary>
-        /// Updates action parameters from UI to model
+        /// Updates action parameters from UI to model.
+        /// Delegated to ParameterPersistenceService.
         /// </summary>
         public void UpdateActionParamsFromUI(DialogNode node, string? scriptName = null)
         {
-            UnifiedLogger.LogApplication(LogLevel.DEBUG,
-                $"UpdateActionParamsFromUI: ENTRY - node '{node.DisplayText}' has {node.ActionParams.Count} existing params");
-
-            node.ActionParams.Clear();
-            var panel = _findControl("ActionsParametersPanel") as StackPanel;
-            if (panel == null)
-            {
-                UnifiedLogger.LogApplication(LogLevel.ERROR,
-                    "UpdateActionParamsFromUI: ActionsParametersPanel NOT FOUND - parameters will be empty!");
-                return;
-            }
-
-            // Get script name from UI if not provided
-            if (string.IsNullOrEmpty(scriptName))
-            {
-                var scriptTextBox = _findControl("ScriptActionTextBox") as TextBox;
-                scriptName = scriptTextBox?.Text;
-            }
-
-            UnifiedLogger.LogApplication(LogLevel.DEBUG,
-                $"UpdateActionParamsFromUI: Found ActionsParametersPanel with {panel.Children.Count} children");
-
-            var autoTrimCheckBox = _findControl("AutoTrimActionsCheckBox") as CheckBox;
-            bool autoTrim = autoTrimCheckBox?.IsChecked ?? true;
-
-            UnifiedLogger.LogApplication(LogLevel.DEBUG, $"UpdateActionParamsFromUI: Auto-Trim = {autoTrim}");
-
-            ProcessParameterPanel(panel, node.ActionParams, autoTrim, scriptName);
-
-            UnifiedLogger.LogApplication(LogLevel.DEBUG,
-                $"UpdateActionParamsFromUI: EXIT - node '{node.DisplayText}' now has {node.ActionParams.Count} params");
-        }
-
-        /// <summary>
-        /// Processes a parameter panel and updates the parameter dictionary
-        /// Issue #287: Fixed to properly find TextBox children regardless of Grid column assignment
-        /// </summary>
-        private void ProcessParameterPanel(StackPanel panel, Dictionary<string, string> paramDict, bool autoTrim, string? scriptName)
-        {
-            UnifiedLogger.LogApplication(LogLevel.DEBUG, $"ProcessParameterPanel: Processing {panel.Children.Count} children");
-
-            foreach (var child in panel.Children)
-            {
-                if (child is Grid paramGrid)
-                {
-                    // Issue #287: Find TextBox children by type, not by index
-                    // Grid children order is not guaranteed to match visual column order
-                    var textBoxes = paramGrid.Children.OfType<TextBox>().ToList();
-
-                    UnifiedLogger.LogApplication(LogLevel.DEBUG,
-                        $"ProcessParameterPanel: Grid has {paramGrid.Children.Count} children, {textBoxes.Count} TextBoxes");
-
-                    if (textBoxes.Count >= 2)
-                    {
-                        // First TextBox is key (column 0), second is value (column 2)
-                        var keyTextBox = textBoxes[0];
-                        var valueTextBox = textBoxes[1];
-
-                        UnifiedLogger.LogApplication(LogLevel.DEBUG,
-                            $"ProcessParameterPanel: keyText='{keyTextBox.Text}', valueText='{valueTextBox.Text}'");
-
-                        if (!string.IsNullOrWhiteSpace(keyTextBox.Text))
-                        {
-                            string key = keyTextBox.Text;
-                            string value = valueTextBox.Text ?? "";
-
-                            // Apply trimming if Auto-Trim is enabled
-                            if (autoTrim)
-                            {
-                                string originalKey = key;
-                                string originalValue = value;
-
-                                key = key.Trim();
-                                value = value.Trim();
-
-                                // Update the UI textboxes to show trimmed values
-                                keyTextBox.Text = key;
-                                valueTextBox.Text = value;
-
-                                // Show visual feedback if text was actually trimmed
-                                if (originalKey != key)
-                                {
-                                    _ = ShowTrimFeedbackAsync(keyTextBox);
-                                }
-                                if (originalValue != value)
-                                {
-                                    _ = ShowTrimFeedbackAsync(valueTextBox);
-                                }
-                            }
-
-                            // Issue #287: Check for duplicate keys
-                            if (paramDict.ContainsKey(key))
-                            {
-                                UnifiedLogger.LogApplication(LogLevel.WARN,
-                                    $"ProcessParameterPanel: Duplicate key '{key}' - overwriting previous value '{paramDict[key]}' with '{value}'");
-                            }
-
-                            paramDict[key] = value;
-                            UnifiedLogger.LogApplication(LogLevel.DEBUG, $"ProcessParameterPanel: Added param '{key}' = '{value}'");
-
-                            // Cache parameter value if script name is known
-                            if (!string.IsNullOrWhiteSpace(scriptName) && !string.IsNullOrWhiteSpace(value))
-                            {
-                                ParameterCacheService.Instance.AddValue(scriptName, key, value);
-                            }
-                        }
-                        else
-                        {
-                            UnifiedLogger.LogApplication(LogLevel.DEBUG, "ProcessParameterPanel: Skipping row with empty key");
-                        }
-                    }
-                    else
-                    {
-                        UnifiedLogger.LogApplication(LogLevel.WARN,
-                            $"ProcessParameterPanel: Grid has {textBoxes.Count} TextBoxes (expected 2+)");
-                    }
-                }
-                else
-                {
-                    UnifiedLogger.LogApplication(LogLevel.WARN, $"ProcessParameterPanel: Child is not Grid, type={child.GetType().Name}");
-                }
-            }
-
-            UnifiedLogger.LogApplication(LogLevel.DEBUG, $"ProcessParameterPanel: Finished with {paramDict.Count} parameters");
-        }
-
-        /// <summary>
-        /// Shows visual feedback when parameter text is trimmed.
-        /// Briefly flashes the TextBox border to indicate successful trim operation.
-        /// Issue #141: Uses theme success color for colorblind accessibility.
-        /// </summary>
-        private async Task ShowTrimFeedbackAsync(TextBox textBox)
-        {
-            // Store original border properties
-            var originalBrush = textBox.BorderBrush;
-            var originalThickness = textBox.BorderThickness;
-
-            try
-            {
-                // Flash success border to indicate trim occurred (theme-aware for accessibility)
-                textBox.BorderBrush = GetSuccessBrush();
-                textBox.BorderThickness = new Thickness(2);
-
-                // Wait briefly for visual feedback
-                await Task.Delay(300);
-
-                // Restore original appearance
-                textBox.BorderBrush = originalBrush;
-                textBox.BorderThickness = originalThickness;
-            }
-            catch (Exception ex)
-            {
-                UnifiedLogger.LogApplication(LogLevel.ERROR, $"ShowTrimFeedback: Error showing visual feedback - {ex.Message}");
-                // Ensure we restore original state even if error occurs
-                textBox.BorderBrush = originalBrush;
-                textBox.BorderThickness = originalThickness;
-            }
+            _persistenceService.UpdateActionParamsFromUI(node, scriptName);
         }
     }
 }
